+++ example/sample-project/manage.py
@@ -1,12 +1,21 @@
 #!/usr/bin/env python
 """Django's command-line utility for administrative tasks."""
+from dotenv import load_dotenv
 import os
 import sys
+
+load_dotenv() 
 
 
 def main():
     """Run administrative tasks."""
-    os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'config.settings')
+    os.environ.setdefault('DJANGO_SETTINGS_MODULE', 
+    'config.settings.development' if os.environ['DEBUG'] else 'config.settings.production'
+    )
+    # if os.environ['DEBUG']:
+    #     os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'config.settings.development')
+    # else: 
+    #     os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'config.settings.production')
     try:
         from django.core.management import execute_from_command_line
     except ImportError as exc: