+++ simple_django_starter/__brick__/{{name.paramCase()}}/config/settings/development.py
@@ -1,0 +1,40 @@
+from .base import *
+
+# Quick-start development settings - unsuitable for production
+# See https://docs.djangoproject.com/en/4.0/howto/deployment/checklist/
+
+# SECURITY WARNING: keep the secret key used in production secret!
+SECRET_KEY = 'django-insecure-yt#=h%x9(qak@x#4==0uap-=w!w22^($=x-1m-#(%ft+(%p7*h'
+
+# SECURITY WARNING: don't run with debug turned on in production!
+DEBUG = True
+
+# Allowed hosts
+ALLOWED_HOSTS = [
+    '*',
+]
+
+
+INSTALLED_APPS += [
+    'debug_toolbar',
+]
+
+MIDDLEWARE += [
+    'debug_toolbar.middleware.DebugToolbarMiddleware', 
+]
+
+DEBUG_TOOLBAR_CONFIG = {
+    'JQUERY_URL': '',
+}
+
+
+# * Database
+# ? https://docs.djangoproject.com/en/4.0/ref/settings/#databases
+DATABASES = {
+    'default': {
+        'ENGINE': 'django.db.backends.sqlite3',
+        'NAME': BASE_DIR / 'db.sqlite3',
+    }
+}
+
+EMAIL_BACKEND = 'django.core.mail.backends.console.EmailBackend'