+++ django_rest_starter/CHANGELOG.md
@@ -1,3 +1,7 @@
 # 0.1.0+1
 
 - Initial release.
+
+# 0.1.0+2
+
+- Readme updated.